requirements-dev.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.21.39
botocore==1.24.39
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-cors-headers==3.10.0
django-filter==21.1
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
environs==9.3.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
marshmallow==3.14.1
oauthlib==3.2.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
django-extensions==3.1.5
